requirements.txt
requests==2.32.5
datetime==5.5
python-dotenv==1.2.1